requirements.txt
asyncpg==0.30.0
discord.py==2.4.0
html2image==2.0.3
Pillow==9.5.0
